refactor(mailbox): splitter send_handlers.rs (718) en pipeline (409) + endpoints (314)
- send_pipeline.rs : 5 helpers pub(crate) (validate/dkim/build/enqueue/dispatch)
- send_endpoints.rs : 4 handlers HTTP (send/undo/schedule/status)
- mod.rs : re-exports mis à jour
- send_queue_worker.rs : imports rebasés
